backend/requirements.txt
Flask==2.2.5
pymongo==4.6.1
scikit-learn==1.2.2
joblib==1.2.0
python-dotenv==1.0.0
pandas==2.0.3


# optional for dev
flask-cors==3.0.10